Refactor location of test files
