+++ test/.resources/favicon.ico
