+++ test/.resources/logo.png
