+++ test/spechelper.js
@@ -30,12 +30,12 @@ exports.TEST_TAG_PREFIX = "cloudinary_npm_test"; // identifies resources created
 exports.TEST_TAG = exports.TEST_TAG_PREFIX + "_" + exports.SUFFIX; // identifies resources created in the current test run with a unique tag
 exports.TEST_ID = exports.TEST_TAG; // identifies resources created in the current test run with a unique id
 exports.UPLOAD_TAGS = [exports.TEST_TAG, exports.TEST_TAG_PREFIX, exports.SDK_TAG];
-exports.IMAGE_FILE = "test/resources/logo.png";
-exports.LARGE_RAW_FILE = "test/resources/TheCompleteWorksOfShakespeare.mobi";
-exports.LARGE_VIDEO = "test/resources/CloudBookStudy-HD.mp4";
-exports.EMPTY_IMAGE = "test/resources/empty.gif";
-exports.RAW_FILE = "test/resources/docx.docx";
-exports.ICON_FILE = "test/resources/favicon.ico";
+exports.IMAGE_FILE = "test/.resources/logo.png";
+exports.LARGE_RAW_FILE = "test/.resources/TheCompleteWorksOfShakespeare.mobi";
+exports.LARGE_VIDEO = "test/.resources/CloudBookStudy-HD.mp4";
+exports.EMPTY_IMAGE = "test/.resources/empty.gif";
+exports.RAW_FILE = "test/.resources/docx.docx";
+exports.ICON_FILE = "test/.resources/favicon.ico";
 exports.VIDEO_URL = "http://res.cloudinary.com/demo/video/upload/dog.mp4";
 exports.IMAGE_URL = "http://res.cloudinary.com/demo/image/upload/sample";
